Move arm dump code to separate C file
arm_core.c is getting far too big and we need to rationalise
it a little before adding the fpa code.
